Fixing Cell-C oddity where user can have 0 minutes, but still have a rand balance, and can therefore still make calls.
I'm not sure how it happened, but it seems like depending on which
system you use to recharge, there is a scenario where you can have
0 minutes but have a rand balance that still allows you to make
calls.

I found the page that returns this rand value and do a simple
calculation to determine the real number of minutes available.

The settings file now has a variable called cellc_voice_per_minute
which is currently set to 0.99.